Implement credentials retrieval
